Fixed update language
