+++ Projects/App/Sources/AppDelegate.swift
@@ -175,16 +175,12 @@ import hGraphQL
             return InterceptingURLSessionClient()
         }
         setupAnalyticsAndTracking()
+        let authenticationStore: AuthenticationStore = globalPresentableStoreContainer.get()
         bag += Localization.Locale.$currentLocale
-            .atOnce()
-            .onValue { locale in
+            .onValue { [weak self] locale in
                 ApplicationState.setPreferredLocale(locale)
                 ApolloClient.acceptLanguageHeader = locale.acceptLanguageHeader
-
-                ApolloClient.initAndRegisterClient()
-                    .always {
-                        self.performUpdateLanguage()
-                    }
+                self?.bag += ApolloClient.initAndRegisterClient()
             }
 
         ApolloClient.bundle = Bundle.main
